keyId: use unsigned 64-bit variables for key ID
Tarsnap "machine num" (or keyfile IDs) are unsigned 64-bit integers.

+++ src/taskmanager.h
@@ -113,7 +113,7 @@ signals:
     //! The tarsnap key ID number.
     //! \param key_filename: the filename.
     //! \param id: the 64-bit unsigned integer representing the key.
-    void keyId(QString key_filename, int id);
+    void keyId(QString key_filename, quint64 id);
     void matchingArchives(QList<ArchivePtr> archives);
 
 private slots:
